Fix hidden circular dependency in watched folders code
