+++ frontend/editor/src/proprietary/components/watchedFolders/WatchedFolderSection.tsx
@@ -16,7 +16,7 @@ import { automationStorage } from "@app/services/automationStorage";
 import {
   WATCHED_FOLDER_VIEW_ID,
   WATCHED_FOLDER_WORKBENCH_ID,
-} from "@app/components/watchedFolders/WatchedFoldersRegistration";
+} from "@app/components/watchedFolders/watchedFolderShared";
 
 export function WatchedFolderSection() {
   const { t } = useTranslation();
